// services/models/AuthResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AuthResponseDTO = {
    /**
     * user's id
     */
    id: string;
    /**
     * user's email
     */
    email?: string;
    /**
     * user's username
     */
    username?: string;
    /**
     * user's picture
     */
    picture?: string;
    /**
     * user's accessToken
     */
    accessToken: string;
    /**
     * user's refreshToken
     */
    refresh_token: string;
};